64-bit computation

+++ src/factorial.rs
@@ -1,4 +1,4 @@
-pub fn factorial(n: u32) -> u32 {
+pub fn factorial(n: u64) -> u64 {
   match n {
     0 => 1,
     _ => n * factorial(n-1),
@@ -6,7 +6,7 @@ pub fn factorial(n: u32) -> u32 {
 }
 
 
-pub fn factorial2(n: u32) -> u32 {
+pub fn factorial2(n: u64) -> u64 {
   (1..n+1).fold(1, |acc, v| acc * v)
 }
